+++ packages/cloud_firestore/cloud_firestore/example/lib/main.dart
@@ -15,11 +15,7 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   if (USE_FIRESTORE_EMULATOR) {
-    FirebaseFirestore.instance.settings = const Settings(
-      host: 'localhost:8080',
-      sslEnabled: false,
-      persistenceEnabled: false,
-    );
+    FirebaseFirestore.instance.useFirestoreEmulator('localhost', 8080);
   }
   runApp(FirestoreExampleApp());
 }