+++ packages/firebase_auth/firebase_auth/example/test_driver/instance_e2e.dart
@@ -692,5 +692,37 @@ void runInstanceTests() {
         expect(credential, isA<PhoneAuthCredential>());
       }, skip: kIsWeb || defaultTargetPlatform != TargetPlatform.android);
     }, skip: defaultTargetPlatform == TargetPlatform.macOS || kIsWeb);
+
+    group('setSettings()', () {
+      test(
+          'throws argument error if phoneNumber & smsCode have not been set simultaneously',
+          () async {
+        String message =
+            "The [smsCode] and the [phoneNumber] must both be either 'null' or a 'String''.";
+        await expectLater(
+            FirebaseAuth.instance.setSettings(phoneNumber: '123456'),
+            throwsA(isA<ArgumentError>()
+                .having((e) => e.message, 'message', contains(message))));
+
+        await expectLater(
+            FirebaseAuth.instance.setSettings(smsCode: '123456'),
+            throwsA(isA<ArgumentError>()
+                .having((e) => e.message, 'message', contains(message))));
+      }, skip: kIsWeb || defaultTargetPlatform != TargetPlatform.android);
+    });
+
+    group('tenantId', () {
+      test('User associated with the tenantId correctly', () async {
+        // tenantId created in the GCP console
+        const String tenantId = 'auth-tenant-test-xukxg';
+        // created User on GCP console associated with the above tenantId
+        final userCredential = await FirebaseAuth.instance
+            .signInWithEmailAndPassword(
+                email: 'test-tenant@email.com', password: 'fake-password');
+
+        expect(userCredential.user.tenantId, tenantId);
+      });
+      // todo(russellwheatley85): get/set tenantId and authenticating user via auth emulator is not possible at the moment.
+    }, skip: true);
   });
 }