+++ packages/firebase_auth/firebase_auth/lib/src/firebase_auth.dart
@@ -73,38 +73,38 @@ class FirebaseAuth extends FirebasePluginPlatform {
 
   /// Changes this instance to point to an Auth emulator running locally.
   ///
-  /// Set the [origin] of the local emulator, such as "http://localhost:9099"
+  /// Set the [host] of the local emulator, such as "localhost"
+  /// Set the [port] of the local emulator, such as "9099" (port 9099 is default for auth package)
   ///
   /// Note: Must be called immediately, prior to accessing auth methods.
   /// Do not use with production credentials as emulator traffic is not encrypted.
-  ///
-  /// Note: auth emulator is not supported for web yet. firebase-js-sdk does not support
-  /// auth.useEmulator until v8.2.4, but FlutterFire does not support firebase-js-sdk v8+ yet
-  Future<void> useEmulator(String origin) async {
-    assert(origin.isNotEmpty);
-    String mappedOrigin = origin;
+  Future<void> useAuthEmulator(String host, int port) async {
+    String mappedHost = host;
 
-    // Android considers localhost as 10.0.2.2 - automatically handle this for users.
-    if (defaultTargetPlatform == TargetPlatform.android && !kIsWeb) {
-      if (mappedOrigin.startsWith('http://localhost')) {
-        mappedOrigin =
-            mappedOrigin.replaceFirst('http://localhost', 'http://10.0.2.2');
-      } else if (mappedOrigin.startsWith('http://127.0.0.1')) {
-        mappedOrigin =
-            mappedOrigin.replaceFirst('http://127.0.0.1', 'http://10.0.2.2');
+    if (!kIsWeb && defaultTargetPlatform == TargetPlatform.android) {
+      if (mappedHost == 'localhost' || mappedHost == '127.0.0.1') {
+        // ignore: avoid_print
+        print('Mapping Auth Emulator host "$mappedHost" to "10.0.2.2".');
+        mappedHost = '10.0.2.2';
       }
     }
 
-    // Native calls take the host and port split out
-    final hostPortRegex = RegExp(r'^http:\/\/([\w\d.]+):(\d+)$');
-    final RegExpMatch? match = hostPortRegex.firstMatch(mappedOrigin);
-    if (match == null) {
-      throw ArgumentError('firebase.auth().useEmulator() origin format error');
-    }
-    // Two non-empty groups in RegExp match - which is null-tested - these are non-null now
-    final String host = match.group(1)!;
-    final int port = int.parse(match.group(2)!);
-    await _delegate.useEmulator(host, port);
+    await _delegate.useAuthEmulator(mappedHost, port);
+  }
+
+  /// The current Auth instance's tenant ID.
+  String? get tenantId {
+    return _delegate.tenantId;
+  }
+
+  /// Set the current Auth instance's tenant ID.
+  ///
+  /// When you set the tenant ID of an Auth instance, all future sign-in/sign-up
+  /// operations will pass this tenant ID and sign in or sign up users to the
+  /// specified tenant project. When set to null, users are signed in to the
+  /// parent project. By default, this is set to `null`.
+  set tenantId(String? tenantId) {
+    _delegate.tenantId = tenantId;
   }
 
   /// Applies a verification code sent to the user by email or other out-of-band
@@ -357,11 +357,16 @@ class FirebaseAuth extends FirebasePluginPlatform {
   Future<void> setSettings({
     bool? appVerificationDisabledForTesting,
     String? userAccessGroup,
+    String? phoneNumber,
+    String? smsCode,
+    bool? forceRecaptchaFlow,
   }) {
     return _delegate.setSettings(
-      appVerificationDisabledForTesting: appVerificationDisabledForTesting,
-      userAccessGroup: userAccessGroup,
-    );
+        appVerificationDisabledForTesting: appVerificationDisabledForTesting,
+        userAccessGroup: userAccessGroup,
+        phoneNumber: phoneNumber,
+        smsCode: smsCode,
+        forceRecaptchaFlow: forceRecaptchaFlow);
   }
 
   /// Changes the current type of persistence on the current Auth instance for