+++ packages/firebase_auth/firebase_auth_web/lib/firebase_auth_web.dart
@@ -118,6 +118,16 @@ class FirebaseAuthWeb extends FirebaseAuthPlatform {
   }
 
   @override
+  String? get tenantId {
+    return _webAuth!.tenantId;
+  }
+
+  @override
+  set tenantId(String? tenantId) {
+    _webAuth!.tenantId = tenantId;
+  }
+
+  @override
   void sendAuthChangesEvent(String appName, UserPlatform? userPlatform) {
     assert(_userChangesListeners[appName] != null);
 
@@ -240,9 +250,13 @@ class FirebaseAuthWeb extends FirebaseAuthPlatform {
   }
 
   @override
-  Future<void> setSettings(
-      {bool? appVerificationDisabledForTesting,
-      String? userAccessGroup}) async {
+  Future<void> setSettings({
+    bool? appVerificationDisabledForTesting,
+    String? userAccessGroup,
+    String? phoneNumber,
+    String? smsCode,
+    bool? forceRecaptchaFlow,
+  }) async {
     _webAuth!.settings.appVerificationDisabledForTesting =
         appVerificationDisabledForTesting;
   }
@@ -358,12 +372,12 @@ class FirebaseAuthWeb extends FirebaseAuthPlatform {
   }
 
   @override
-  Future<void> useEmulator(String host, int port) async {
+  Future<void> useAuthEmulator(String host, int port) async {
     try {
       // The generic platform interface is with host and port split to
       // centralize logic between android/ios native, but web takes the
       // origin as a single string
-      _webAuth!.useEmulator('http://$host:$port');
+      _webAuth!.useAuthEmulator('http://$host:$port');
     } catch (e) {
       throw getFirebaseAuthException(e);
     }