+++ packages/firebase_crashlytics/firebase_crashlytics/android/src/main/java/io/flutter/plugins/firebase/crashlytics/FlutterFirebaseCrashlyticsPlugin.java
@@ -6,13 +6,16 @@ package io.flutter.plugins.firebase.crashlytics;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.analytics.connector.AnalyticsConnector;
 import com.google.firebase.crashlytics.FirebaseCrashlytics;
+import com.google.firebase.crashlytics.internal.analytics.CrashlyticsOriginAnalyticsEventLogger;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
@@ -118,6 +121,7 @@ public class FlutterFirebaseCrashlyticsPlugin
           final String reason = (String) arguments.get(Constants.REASON);
           final String information =
               (String) Objects.requireNonNull(arguments.get(Constants.INFORMATION));
+          final boolean fatal = (boolean) arguments.get(Constants.FATAL);
 
           Exception exception;
           if (reason != null) {
@@ -128,6 +132,22 @@ public class FlutterFirebaseCrashlyticsPlugin
           } else {
             exception = new FlutterError(dartExceptionMessage);
           }
+
+          if (fatal) {
+            AnalyticsConnector connector = FirebaseApp.getInstance().get(AnalyticsConnector.class);
+            CrashlyticsOriginAnalyticsEventLogger analyticsEventLogger =
+                new CrashlyticsOriginAnalyticsEventLogger(connector);
+
+            Bundle params = new Bundle();
+            long unixTime = System.currentTimeMillis() / 1000;
+
+            params.putInt(Constants.FATAL, 1);
+            params.putLong(Constants.TIMESTAMP, unixTime);
+
+            crashlytics.setCustomKey(Constants.CRASH_EVENT_KEY, unixTime);
+            analyticsEventLogger.logEvent(Constants.FIREBASE_APPLICATION_EXCEPTION, params);
+          }
+
           crashlytics.setCustomKey(Constants.FLUTTER_ERROR_EXCEPTION, dartExceptionMessage);
 
           final List<StackTraceElement> elements = new ArrayList<>();
@@ -293,6 +313,7 @@ public class FlutterFirebaseCrashlyticsPlugin
   }
 
   // TODO remove once Crashlytics public API supports isCrashlyticsCollectionEnabled
+
   /**
    * Firebase Crashlytics SDK doesn't provide a way to read current enabled status. So we read it
    * ourselves from shared preferences instead.