+++ packages/firebase_crashlytics/firebase_crashlytics/ios/Classes/FLTFirebaseCrashlyticsPlugin.m
@@ -5,6 +5,14 @@
 #import "FLTFirebaseCrashlyticsPlugin.h"
 
 #import <Firebase/Firebase.h>
+
+#if TARGET_OS_OSX
+// macOS platform does not support analytics
+#else
+#import <FirebaseCrashlytics/FIRAnalyticsInterop.h>
+#import <FirebaseCrashlytics/FIRCLSAnalyticsManager.h>
+#endif
+
 #import <firebase_core/FLTFirebasePluginRegistry.h>
 
 NSString *const kFLTFirebaseCrashlyticsChannelName = @"plugins.flutter.io/firebase_crashlytics";
@@ -17,6 +25,7 @@ NSString *const kCrashlyticsArgumentReason = @"reason";
 NSString *const kCrashlyticsArgumentIdentifier = @"identifier";
 NSString *const kCrashlyticsArgumentKey = @"key";
 NSString *const kCrashlyticsArgumentValue = @"value";
+NSString *const kCrashlyticsArgumentFatal = @"fatal";
 
 NSString *const kCrashlyticsArgumentFile = @"file";
 NSString *const kCrashlyticsArgumentLine = @"line";
@@ -25,6 +34,18 @@ NSString *const kCrashlyticsArgumentMethod = @"method";
 NSString *const kCrashlyticsArgumentEnabled = @"enabled";
 NSString *const kCrashlyticsArgumentUnsentReports = @"unsentReports";
 NSString *const kCrashlyticsArgumentDidCrashOnPreviousExecution = @"didCrashOnPreviousExecution";
+
+#if TARGET_OS_OSX
+// macOS platform does not support analytics
+#else
+@interface FIRCLSAnalyticsManager ()
+@property(nonatomic, strong) id<FIRAnalyticsInterop> analytics;
+@end
+
+@interface FIRCrashlytics ()
+@property(nonatomic, strong) FIRCLSAnalyticsManager *analyticsManager;
+@end
+#endif
 
 @interface FLTFirebaseCrashlyticsPlugin ()
 @end
@@ -104,6 +125,7 @@ NSString *const kCrashlyticsArgumentDidCrashOnPreviousExecution = @"didCrashOnPr
   NSString *information = arguments[kCrashlyticsArgumentInformation];
   NSString *dartExceptionMessage = arguments[kCrashlyticsArgumentException];
   NSArray *errorElements = arguments[kCrashlyticsArgumentStackTraceElements];
+  BOOL fatal = [arguments[kCrashlyticsArgumentFatal] boolValue];
 
   // Log additional information so it's captured on the Firebase Crashlytics dashboard.
   if ([information length] != 0) {
@@ -123,6 +145,18 @@ NSString *const kCrashlyticsArgumentDidCrashOnPreviousExecution = @"didCrashOnPr
                                           forKey:@"flutter_error_reason"];
   } else {
     reason = dartExceptionMessage;
+  }
+
+  if (fatal) {
+    NSTimeInterval timeInterval = [NSDate date].timeIntervalSince1970;
+    [[FIRCrashlytics crashlytics] setCustomValue:@(llrint(timeInterval))
+                                          forKey:@"com.firebase.crashlytics.flutter.fatal"];
+#if TARGET_OS_OSX
+    // macOS platform does not support analytics
+#else
+    id<FIRAnalyticsInterop> analytics = [[FIRCrashlytics crashlytics].analyticsManager analytics];
+    [FIRCLSAnalyticsManager logCrashWithTimeStamp:timeInterval toAnalytics:analytics];
+#endif
   }
 
   // Log additional custom value to match Android.
@@ -189,8 +223,12 @@ NSString *const kCrashlyticsArgumentDidCrashOnPreviousExecution = @"didCrashOnPr
 #pragma mark - Utilities
 
 - (FIRStackFrame *)generateFrame:(NSDictionary *)errorElement {
+  NSString *methodName = [errorElement valueForKey:kCrashlyticsArgumentMethod];
+  NSString *className = [errorElement valueForKey:@"class"];
+  NSString *symbol = [NSString stringWithFormat:@"%@.%@", className, methodName];
+
   FIRStackFrame *frame = [FIRStackFrame
-      stackFrameWithSymbol:[errorElement valueForKey:kCrashlyticsArgumentMethod]
+      stackFrameWithSymbol:symbol
                       file:[errorElement valueForKey:kCrashlyticsArgumentFile]
                       line:[[errorElement valueForKey:kCrashlyticsArgumentLine] intValue]];
   return frame;