+++ packages/firebase_app_check/firebase_app_check_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.1+3
+
+ - Update a dependency to the latest release.
+
+## 0.0.1+2
+
+ - Update a dependency to the latest release.
+
 ## 0.0.1+1
 
  - Update a dependency to the latest release.