+++ packages/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0+7
+
+ - Update a dependency to the latest release.
+
+## 0.5.0+6
+
+ - Update a dependency to the latest release.
+
 ## 0.5.0+5
 
  - Update a dependency to the latest release.