+++ packages/firebase_messaging/firebase_messaging/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 10.0.4
+
+ - **DOCS**: update web example in line with flutter 2.2.0 generated `index.html` (#6398).
+ - **CHORE**: update v2 embedding support (#6506).
+ - **CHORE**: rm deprecated jcenter repository (#6431).
+
+## 10.0.3
+
+ - **DOCS**: update web example in line with flutter 2.2.0 generated `index.html` (#6398).
+ - **CHORE**: rm deprecated jcenter repository (#6431).
+
+## 10.0.2
+
+ - Update a dependency to the latest release.
+
 ## 10.0.1
 
  - **FIX**: Fix FirebaseMessaging.onMessage and onMessageOpenedApp potentially throwing (#6093).