+++ packages/firebase_remote_config/firebase_remote_config/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.10.0+3
+
+ - Update a dependency to the latest release.
+
+## 0.10.0+2
+
+ - Update a dependency to the latest release.
+
 ## 0.10.0+1
 
  - Update a dependency to the latest release.