+++ packages/firebase_storage/firebase_storage_web/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 3.0.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: use<product>Emulator(host, port) API update (#6439).
+
+## 2.0.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: useStorageEmulator(host, port) API update.
+
+## 1.1.2
+
+ - Update a dependency to the latest release.
+
 ## 1.1.1
 
  - Update a dependency to the latest release.